Added noop module feature which allows for a module to only be the calling of a query code.   The results can then be shared amongst following modules.   Almost a global data definition without all the additonal code changes that a global one would require.

+++ libexec/begblock.py
@@ -73,41 +73,44 @@ def begblock(argv):
             if modname not in config[pfwdefs.SW_MODULESECT]:
                 miscutils.fwdie("Error: Could not find module description for module %s\n" % \
                                 (modname), pfwdefs.PF_EXIT_FAILURE)
+            moddict = config[pfwdefs.SW_MODULESECT][modname]
 
-            (infsect, outfsect) = pfwblock.get_datasect_types(config, modname)
             runqueries(config, configfile, modname, modules_prev_in_list)
             pfwblock.read_master_lists(config, modname, masterdata, modules_prev_in_list)
 
+            (infsect, outfsect) = pfwblock.get_datasect_types(config, modname)
             pfwblock.fix_master_lists(config, modname, masterdata, outfsect)
-            pfwblock.create_fullnames(config, modname, masterdata)
-            if miscutils.fwdebug_check(9, 'PFWBLOCK_DEBUG') and modname in masterdata:
-                with open('%s-masterdata.txt' % modname, 'w') as fh:
-                    miscutils.pretty_print_dict(masterdata[modname], fh)
+
+            if pfwdefs.PF_NOOP not in moddict or not miscutils.convertBool(moddict[pfwdefs.PF_NOOP]):
+                pfwblock.create_fullnames(config, modname, masterdata)
+                if miscutils.fwdebug_check(9, 'PFWBLOCK_DEBUG') and modname in masterdata:
+                    with open('%s-masterdata.txt' % modname, 'w') as fh:
+                        miscutils.pretty_print_dict(masterdata[modname], fh)
             
-            pfwblock.add_file_metadata(config, modname)
-            sublists = pfwblock.create_sublists(config, modname, masterdata)
-            if sublists is not None:
-                if miscutils.fwdebug_check(3, 'PFWBLOCK_DEBUG'):
-                    miscutils.fwdebug_print("sublists.keys() = %s" % (sublists.keys()))
-            loopvals = pfwblock.get_wrapper_loopvals(config, modname)
-            wrapinst = pfwblock.create_wrapper_inst(config, modname, loopvals)
-            wcnt = 1
-            for winst in wrapinst.values():
-                stime = time.time()
-                if miscutils.fwdebug_check(6, 'PFWBLOCK_DEBUG'):
-                    miscutils.fwdebug_print("winst %d - BEG" % wcnt)
-                pfwblock.assign_data_wrapper_inst(config, modname, winst, masterdata,
-                                                  sublists, infsect, outfsect)
-                modinputs, modoutputs = pfwblock.finish_wrapper_inst(config, modname, winst,
-                                                                     outfsect)
-                inputfiles.extend(modinputs)
-                outputfiles.extend(modoutputs)
-                pfwblock.create_module_wrapper_wcl(config, modname, winst)
-                pfwblock.divide_into_jobs(config, modname, winst, joblist, parlist)
-                etime = time.time()
-                if miscutils.fwdebug_check(6, 'PFWBLOCK_DEBUG'):
-                    miscutils.fwdebug_print("winst %d - %s - END" % (wcnt, etime-stime))
-                wcnt += 1
+                pfwblock.add_file_metadata(config, modname)
+                sublists = pfwblock.create_sublists(config, modname, masterdata)
+                if sublists is not None:
+                    if miscutils.fwdebug_check(3, 'PFWBLOCK_DEBUG'):
+                        miscutils.fwdebug_print("sublists.keys() = %s" % (sublists.keys()))
+                loopvals = pfwblock.get_wrapper_loopvals(config, modname)
+                wrapinst = pfwblock.create_wrapper_inst(config, modname, loopvals)
+                wcnt = 1
+                for winst in wrapinst.values():
+                    stime = time.time()
+                    if miscutils.fwdebug_check(6, 'PFWBLOCK_DEBUG'):
+                        miscutils.fwdebug_print("winst %d - BEG" % wcnt)
+                    pfwblock.assign_data_wrapper_inst(config, modname, winst, masterdata,
+                                                      sublists, infsect, outfsect)
+                    modinputs, modoutputs = pfwblock.finish_wrapper_inst(config, modname, winst,
+                                                                         outfsect)
+                    inputfiles.extend(modinputs)
+                    outputfiles.extend(modoutputs)
+                    pfwblock.create_module_wrapper_wcl(config, modname, winst)
+                    pfwblock.divide_into_jobs(config, modname, winst, joblist, parlist)
+                    etime = time.time()
+                    if miscutils.fwdebug_check(6, 'PFWBLOCK_DEBUG'):
+                        miscutils.fwdebug_print("winst %d - %s - END" % (wcnt, etime-stime))
+                    wcnt += 1
             modules_prev_in_list[modname] = True
 
             if miscutils.fwdebug_check(9, 'PFWBLOCK_DEBUG') and modname in masterdata:
